Move the the custom code to a trait.
