+++ .gitignore
@@ -1,11 +1,2 @@
-Custom/SiteCloneCustom.php
-
-# Created by .ignore support plugin (hsz.mobi)
-### Composer template
-composer.phar
 /vendor/
-
-# Commit your application's lock file http://getcomposer.org/doc/01-basic-usage.md#composer-lock-the-lock-file
-# You may choose to ignore a library lock file http://getcomposer.org/doc/02-libraries.md#lock-file
-# composer.lock
