+++ Custom/SiteCloneCustomTrait.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SiteClone\Custom;
+
+trait SiteCloneCustomTrait {
+
+  // Use this file to add custom code. See README.md for details.
+
+}
